// src/api/document/request/data/check.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface Check {
  x: number;
  y: number;
  width: number;
  height: number;
  subtype: string;
  page_number: number;
}