// src/api/document/request/data/deactivateElement.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface DeactivateElement {
  type: string;
  unique_id: string;
}