// src/api/document/request/data/text.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface Text {
  x: number;
  y: number;
  size: number;
  width: number;
  height: number;
  subtype: string;
  page_number: number;
  data: string;
  font: string;
  line_height: number;
  field_id?: string;
}