// src/api/documentInvite/request/data/ccStep.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface CcStep {
  name: string;
  email: string;
  step: number;
}