// src/api/documentGroupInvite/request/data/completionEmail.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface CompletionEmail {
  email?: string;
  disable_document_attachment?: number;
  subject?: string;
  message?: string;
}